floatをdoubleにすべて変換した

+++ measure/Measure.cpp
@@ -5,7 +5,7 @@ Measure::Measure()
     mLpf = new LowPassFilter();
 }
 
-float Measure::getValue()
+double Measure::getValue()
 {
     return 0;
 }
