+++ walker/SimpleWalker.h
@@ -26,7 +26,7 @@ public:
 
     virtual void run();
     void setCommand(int forward, int turn);
-    void setCommandV(float forward, int turn);
+    void setCommandV(double forward, int turn);
     void resetParam();
 
     virtual void init();
@@ -34,7 +34,7 @@ public:
 
 protected:
 
-    float mForward;
+    double mForward;
     int mTurn;
 
     bool mBreake_flag;
